// src/types/match.ts
export interface MatchedPets {
  match_id: string;
  status: string;
  match_created_at: string;
  pet1_id: string;
  pet1_name: string;
  pet1_breed: string;
  pet1_age: number;
  pet1_owner_name: string;
  pet2_id: string;
  pet2_name: string;
  pet2_breed: string;
  pet2_age: number;
  pet2_owner_name: string;
}